refactor(gemma4): split multimodal support into dedicated module
